pyproject.toml
[project]
name = "tts-generator"
version = "0.1.0"
description = "Markdown 및 텍스트 기반 웹 TTS 서비스"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn>=0.30.0,<1.0.0",
  "python-multipart>=0.0.9,<1.0.0",
]

[project.optional-dependencies]
melo = [
  "setuptools<81",
  "python-mecab-ko>=1.3.7",
  "python-mecab-ko-dic>=2.1.1.post2",
  "melotts @ git+https://github.com/myshell-ai/MeloTTS.git",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
